Modified Prelim/load_behavior_data_SAT() to load MG data for Quincy and Seymour. Also added fxn plot_Pcorr_vs_dir_MG for the memory-guided task.

+++ RespTime/plot_CDF_RT_SAT.m
@@ -28,7 +28,7 @@ for kk = 1:NUM_SESSION
   idx_errdir = info(kk).err_dir;
   idx_corr = ~(info(kk).err_dir | info(kk).err_hold | info(kk).err_nosacc);
   
-  RT = moves(kk).resptime;
+  RT = double(moves(kk).resptime);
   
   RT_corr_A(kk,:) = quantile(RT(idx_corr & idx_acc), QUANTILE);
   RT_errdir_A(kk,:) = quantile(RT(idx_errdir & idx_acc), QUANTILE);
